### 2024-06-16 #85 @disha101003
[Re] Bio-inspired replay for transformers in Continual Learning
**Original article:** https://arxiv.org/abs/2305.04769

**Metadata URL:**

**Scientific domain:** Computer Vision and Pattern Recognition
