### 2024-10-11 #85 @rougier
I'll edit this submission then.

@zhywan @MollyZhang @arianesasso Can you review this submission ?
Bio-inspired replay for transformers in Continual Learning
