### 2024-11-04 138f485
Fixed README.md and added versions in requirements.txt
