+++ requirements.txt
@@ -1,10 +1,52 @@
-mlflow
-matplotlib
-numpy
-einops
-torch
-torchsummary
-tqdm
-torchvision
-dagshub
-ipython
+asttokens==2.4.1
+contourpy==1.3.0
+cycler==0.12.1
+decorator==5.1.1
+einops==0.8.0
+executing==2.1.0
+filelock==3.16.1
+fonttools==4.54.1
+fsspec==2024.10.0
+ipython==8.29.0
+jedi==0.19.1
+Jinja2==3.1.4
+kiwisolver==1.4.7
+MarkupSafe==3.0.2
+matplotlib==3.9.2
+matplotlib-inline==0.1.7
+mpmath==1.3.0
+networkx==3.4.2
+numpy==2.1.3
+nvidia-cublas-cu12==12.4.5.8
+nvidia-cuda-cupti-cu12==12.4.127
+nvidia-cuda-nvrtc-cu12==12.4.127
+nvidia-cuda-runtime-cu12==12.4.127
+nvidia-cudnn-cu12==9.1.0.70
+nvidia-cufft-cu12==11.2.1.3
+nvidia-curand-cu12==10.3.5.147
+nvidia-cusolver-cu12==11.6.1.9
+nvidia-cusparse-cu12==12.3.1.170
+nvidia-nccl-cu12==2.21.5
+nvidia-nvjitlink-cu12==12.4.127
+nvidia-nvtx-cu12==12.4.127
+packaging==24.1
+parso==0.8.4
+pexpect==4.9.0
+pillow==11.0.0
+prompt_toolkit==3.0.48
+ptyprocess==0.7.0
+pure_eval==0.2.3
+Pygments==2.18.0
+pyparsing==3.2.0
+python-dateutil==2.9.0.post0
+six==1.16.0
+stack-data==0.6.3
+sympy==1.13.1
+torch==2.5.1
+torchsummary==1.5.1
+torchvision==0.20.1
+tqdm==4.66.6
+traitlets==5.14.3
+triton==3.1.0
+typing_extensions==4.12.2
+wcwidth==0.2.13
