+++ src/models/test.py
@@ -44,10 +44,14 @@ if __name__ == '__main__':
     testloader = test_data()
     criterion = nn.CrossEntropyLoss()
     model_g, model_f_w, model_f_s = get_models()
-
-    model_g.load_state_dict(torch.load(const.MODEL_DIR/'model_g'))
-    model_f_w.load_state_dict(torch.load(const.MODEL_DIR/'model_f_w'))
-    model_f_s.load_state_dict(torch.load(const.MODEL_DIR/'model_f_s'))
+    if const.PRETRAINED == True:
+        model_g.load_state_dict(torch.load(const.model_g_path))
+        model_f_w.load_state_dict(torch.load(const.model_fw_path))
+        model_f_s.load_state_dict(torch.load(const.model_fs_path))
+    else:
+        model_g.load_state_dict(torch.load(const.MODEL_DIR/'model_g'))
+        model_f_w.load_state_dict(torch.load(const.MODEL_DIR/'model_f_w'))
+        model_f_s.load_state_dict(torch.load(const.MODEL_DIR/'model_f_s'))
     test(criterion,testloader, model_g, model_f_w, model_f_s )
 
 
