### 2024-11-05 82928a9
Added instructions for using pretrained models in README.md
