### 2024-11-05 69b3b35
Added instructions for using pretrained models in README.md
