+++ src/const.py
@@ -72,6 +72,6 @@ PRETRAINED = True
 # PRETRAINED_BASE_PATH = BASE_DIR/'models'/'pretrained_models_CiFAR10' for CiFAR10
 # PRETRAINED_BASE_PATH = BASE_DIR/'models'/'pretrained_models_CiFAR100'for CiFAR100
 PRETRAINED_BASE_PATH = BASE_DIR/'models'/'pretrained_models_CiFAR100'
-model_fs_path = PRETRAINED_BASE_PATH / 'cifar10t4_fs'
-model_fw_path = PRETRAINED_BASE_PATH / 'cifar10t4_fw'
-model_g_path = PRETRAINED_BASE_PATH / 'cifar10t4_g'
+model_fs_path = PRETRAINED_BASE_PATH / 'cifar100ft4_fs.pth'
+model_fw_path = PRETRAINED_BASE_PATH / 'cifar100ft4_fw.pth'
+model_g_path = PRETRAINED_BASE_PATH / 'cifar100ft4_g.pth'
