+++ src/models/train_task.py
@@ -1,153 +1,196 @@
 #!/usr/bin/env python3
-from tqdm import tqdm, trange
+from tqdm import tqdm
 
 from src import const
 from src.dataset import train_data
-import torch.nn.functional as F
 import time
 import torch
 from torch import nn
 import numpy as np
 import torch.optim as optim
 import os
-import pickle
 
 from src.models.arch import get_models, D_buffer
 
 
-def train_task(task_index, trainloader, sem_mem, model_g, model_f_w, model_f_s, criterion, optimizer):
+def train_task(
+        task_index,
+        trainloader,
+        sem_mem,
+        model_g,
+        model_f_w,
+        model_f_s,
+        criterion,
+        optimizer):
     c = 0
     # training loop of BiRT
-    start = int(time.time()/60) # task training start time
+    start = int(time.time() / 60)  # task training start time
     for epoch in range(const.epochs):
         train_loss = 0.0
         task_sem_mem_list = []
 
-        for batch_idx, batch in enumerate(tqdm((trainloader), desc=f"Epoch {epoch + 1} / {const.epochs} in training", leave=False)):
+        for batch_idx, batch in enumerate(tqdm(
+                (trainloader), desc=f"Epoch {epoch + 1} / \
+                    {const.epochs} in training", leave=False)):
             x, y = batch
             x, y = x.to(const.device), y.to(const.device)
-            y_hat_g,_ = model_g(x,c)                      # get output from g()
-            y_hat_temp_copy = y_hat_g.detach().clone()  # create a detached copy to be processed further
+            # get output from g()
+            y_hat_g, _ = model_g(x, c)
+            # create a detached copy to be processed further
+            y_hat_temp_copy = y_hat_g.detach().clone()
 
             # store output from g() along with labels in episodic memory
             y_hat_temp_mem = y_hat_temp_copy.clone()
             y_mem = y
             task_sem_mem_list.append((y_hat_temp_mem, y_mem))
 
-            # controls sampled from a normal distribution to control different noises introduced in BiRT
-            alpha_t_comp, alpha_a_comp, alpha_s_comp, alpha_e_comp = np.random.uniform(0,1,4)
+            # controls sampled from a normal distribution to control different
+            # noises introduced in BiRT
+            alpha_t_comp, alpha_a_comp, alpha_s_comp, alpha_e_comp = \
+                np.random.uniform(0, 1, 4)
 
             # sampling mini batch from episodic memory
-            if (sem_mem.is_empty() == False):
+            if not sem_mem.is_empty():
                 r, r_y = sem_mem.get_batch()
                 r = r.to((const.device))
                 r_y = r_y.to(const.device)
 
                 # implementing label noise
-                if(alpha_t_comp < const.alpha_t):
-                    num_change = int(const.percentage_change / 100 * const.batch_size)
-                    indices_change_r_y = torch.randperm(len(y))[:num_change].to(const.device)
-                    r_y_changed = torch.randint(0,const.num_classes,(num_change,)).to(const.device)
+                if (alpha_t_comp < const.alpha_t):
+                    num_change = int(
+                        const.percentage_change / 100 * const.batch_size)
+                    indices_change_r_y = torch.randperm(
+                        len(y))[:num_change].to(const.device)
+                    r_y_changed = torch.randint(
+                        0, const.num_classes, (num_change,)).to(
+                        const.device)
                     r_y[indices_change_r_y] = r_y_changed
 
[…]
