### 2024-11-06 1ebb641
Added doc strings
