### 2024-11-07 873969a
Added support for running all pretrained models
