+++ src/const.py
@@ -5,18 +5,18 @@ import torch
 
 BASE_DIR = Path(__file__).parent.parent
 DATA_DIR = BASE_DIR / 'data'
+# DATASET = 'CiFAR10' for CIFAR10
+# DATASET = 'CiFAR100' for CIFAR10
+DATASET = 'CiFAR10'
 # ViT_CIFAR10_data when training for CiFAR10
 # ViT_CIFAR10_data when training for CiFAR100
-DATA_TEST_DIR = DATA_DIR / 'ViT_CIFAR100_data'
+DATA_TEST_DIR = DATA_DIR / f'ViT_{DATASET}_data'
 # MODEL_DIR = BASE_DIR / 'models'/'models_CiFAR10' for CiFAR10
 # MODEL_DIR = BASE_DIR / 'models'/'models_CiFAR100' for CiFAR100
-MODEL_DIR = BASE_DIR / 'models' / 'models_CiFAR100'
+MODEL_DIR = BASE_DIR / 'models' / f'models_{DATASET}'
 # MEMORY_DIR = BASE_DIR / 'memory'/'memory_CiFAR10' for CiFAR10
 # MEMORY_DIR = BASE_DIR / 'memory'/'memory_CiFAR100' for CiFAR100
-MEMORY_DIR = BASE_DIR / 'memory' / 'memory_CiFAR100'
-# DATASET = 'CIFAR10' for CIFAR10
-# DATASET = 'CIFAR100' for CIFAR10
-DATASET = 'CIFAR100'
+MEMORY_DIR = BASE_DIR / 'memory' / f'memory_{DATASET}'
 # SUBSET = 0 when using the entire dataset
 # SUBSET = 1 when using a subset of dataset to train the model on
 SUBSET = 0
@@ -25,6 +25,9 @@ SUBSET = 0
 NUM_SUBSET_IMAGES = 1000
 # Number of Images to Fine Tune
 FINE_TUNE_SIZE = 1000
+# NUM_CLASSES = 10 for CiFAR10
+# NUM_CLASSES = 100 for CiFAR100
+NUM_CLASSES = 10
 config = {
     "patch_size": 4,
     "hidden_size": 48,
@@ -44,7 +47,7 @@ config = {
 # defining dataset/training specific parameters
 base_lr = 5e-4
 weight_decay = 1e-6
-num_classes = 100  # num_classes of CIFAR100 / CIFAR10
+num_classes = NUM_CLASSES  # num_classes of CIFAR100 / CIFAR10
 accum_iter = 4
 tasks = 5  # num of tasks
 epochs = 200 # num of epochs for training 
@@ -73,14 +76,3 @@ sem_mem_length = 500
 task_num = 1
 
 device = 'cuda:0' if torch.cuda.is_available() else 'cpu'  # choosing device
-
-# For Testing with  pretrained models only
-PRETRAINED = True
-# PRETRAINED_BASE_PATH = BASE_DIR/'models'/'pretrained_models_CiFAR10' for
-# CiFAR10
-# PRETRAINED_BASE_PATH = BASE_DIR/'models'/'pretrained_models_CiFAR100'for
-# CiFAR100
-PRETRAINED_BASE_PATH = BASE_DIR / 'models' / 'pretrained_models_CiFAR100'
-model_fs_path = PRETRAINED_BASE_PATH / 'cifar100ft4_fs.pth'
-model_fw_path = PRETRAINED_BASE_PATH / 'cifar100ft4_fw.pth'
-model_g_path = PRETRAINED_BASE_PATH / 'cifar100ft4_g.pth'
