### 2024-11-07 c4e4660
Added support for testing pretrained models
