+++ src/models/test_pretrained_models.py
@@ -3,7 +3,7 @@ from src import const
 from src.dataset import test_data
 import argparse
 from pathlib import Path
-import torch 
+import torch
 from torch import nn
 
 # For Testing with  pretrained models only
@@ -11,16 +11,26 @@ from torch import nn
 # CiFAR10
 # PRETRAINED_BASE_PATH = BASE_DIR/'models'/'pretrained_models_CiFAR100'for
 # CiFAR100
-PRETRAINED_BASE_PATH = const.BASE_DIR / 'models' / 'pretrained_models_CiFAR10'
+PRETRAINED_BASE_PATH = const.BASE_DIR / \
+    'models' / f'pretrained_models_{const.DATASET}'
+
 
 def main():
     # Set up argument parser
     parser = argparse.ArgumentParser()
-    
-    parser.add_argument('--model_fw_path', type=Path, default=PRETRAINED_BASE_PATH / 'cifar10ft4_fw.pth',
-                        help='Path to the model file for FW (default: models/pretrained_models_CiFAR10/cifar10ft4_fw.pth)')
-    parser.add_argument('--model_g_path', type=Path, default=PRETRAINED_BASE_PATH / 'cifar100ft4_g.pth',
-                        help='Path to the model file for G (default: models/pretrained_models_CiFAR10/cifar10ft4_g.pth)')
+
+    parser.add_argument(
+        '--model_fw_path',
+        type=Path,
+        default=PRETRAINED_BASE_PATH /
+        f'{const.DATASET.lower()}ft4_fw.pth',
+        help='Path to the model file for fw')
+    parser.add_argument(
+        '--model_g_path',
+        type=Path,
+        default=PRETRAINED_BASE_PATH /
+        f'{const.DATASET.lower()}ft4_g.pth',
+        help='Path to the model file for g')
 
     # Parse arguments
     args = parser.parse_args()
@@ -31,6 +41,6 @@ def main():
     model_f_w = torch.load(args.model_fw_path, weights_only=False)
     test(criterion, testloader, model_g, model_f_w)
 
+
 if __name__ == "__main__":
     main()
-
