### 2024-11-07 6cf46e1
Added bash script for training and testing
