### 2025-02-17 #85 @rougier
Since we're very late on review and base on through review by @gdetor, I propose to accept the article. Congratulations. Can you point me to tour article source repository since I'll need to produce the final PDF.
