### 2025-02-17 #85 @disha101003
Hi,
@rougier Thank you so much!
This is the github repository for the replication study: https://github.com/disha101003/ReScience
This is the link to the source directory for the replication study: https://github.com/disha101003/ReScience/tree/master/src
This is the link to the README for the replication study which explains the directory structure of the repository: https://github.com/disha101003/ReScience/blob/master/README.md

This is the link to the original article: https://arxiv.org/abs/2305.04769
This is the link to the original article code repository: https://github.com/NeurAI-Lab/BiRT

Please let me know if you need anything else.
