### 2025-02-17 #85 @rougier
Thank you for all these links. But the one I need to the repository for you article sources (I need to edit editor / reviewers and generate a new PDF with the DOI).

In the meantime, can you save yopur github repository on [software heritage](https://archive.softwareheritage.org/save/) and get a swid (this is a permanent id on the archive, it's better than Zenodo).
