### 2025-02-18 #85 @rougier
I removed the link from your comment because it is a security issue. Anybody that click on the link can edit/erase your article. Please make sure to make a copy or your overleaf repo just in case and then send me by email an invitation to edit your overleaf.
