### 2025-02-26 #85 @rougier
Can you update the metadata.yaml file in your overleaf directory and send it back here? You'll miss some information but I'll complete them.
