### 2025-03-01 #85 @rougier
Perfect. Here is the **sandbox** version (not final) of your article: https://sandbox.zenodo.org/records/176429
Can you check eveyrthing's correct. I'm waiting for you go and I'll publish the actual version.
