### 2025-03-04 #85 @rougier
It's online at https://zenodo.org/record/14964875 and will soon appear on the rescience website.
Congratulations!
